Snapshot 73.0.3683.119

+++ src/base/i18n/time_formatting.cc
@@ -221,10 +221,10 @@ bool TimeDurationFormatWithSeconds(const TimeDelta time,
                                    string16* out) {
   DCHECK(out);
   UErrorCode status = U_ZERO_ERROR;
-  const int64_t total_seconds = static_cast<int>(time.InSecondsF() + 0.5);
-  const int hours = total_seconds / 3600;
-  const int minutes = (total_seconds - hours * 3600) / 60;
-  const int seconds = total_seconds % 60;
+  const int64_t total_seconds = static_cast<int64_t>(time.InSecondsF() + 0.5);
+  const int64_t hours = total_seconds / 3600;
+  const int64_t minutes = (total_seconds - hours * 3600) / 60;
+  const int64_t seconds = total_seconds % 60;
   UMeasureFormatWidth u_width = DurationWidthToMeasureWidth(width);
 
   const icu::Measure measures[] = {
