+++ src/gpu/command_buffer/client/ring_buffer.h
@@ -22,7 +22,7 @@ class CommandBufferHelper;
 // allocations must not be kept past new allocations.
 class GPU_EXPORT RingBuffer {
  public:
-  typedef unsigned int Offset;
+  typedef uint32_t Offset;
 
   // Creates a RingBuffer.
   // Parameters:
@@ -31,9 +31,9 @@ class GPU_EXPORT RingBuffer {
   //   size: The size of the buffer in bytes.
   //   helper: A CommandBufferHelper for dealing with tokens.
   //   base: The physical address that corresponds to base_offset.
-  RingBuffer(unsigned int alignment,
+  RingBuffer(uint32_t alignment,
              Offset base_offset,
-             size_t size,
+             uint32_t size,
              CommandBufferHelper* helper,
              void* base);
 
@@ -51,7 +51,7 @@ class GPU_EXPORT RingBuffer {
   //
   // Returns:
   //   the pointer to the allocated memory block.
-  void* Alloc(unsigned int size);
+  void* Alloc(uint32_t size);
 
   // Frees a block of memory, pending the passage of a token. That memory won't
   // be re-allocated until the token has passed through the command stream.
@@ -61,7 +61,7 @@ class GPU_EXPORT RingBuffer {
   // Parameters:
   //   pointer: the pointer to the memory block to free.
   //   token: the token value to wait for before re-using the memory.
-  void FreePendingToken(void* pointer, unsigned int token);
+  void FreePendingToken(void* pointer, uint32_t token);
 
   // Discards a block within the ring buffer.
   //
@@ -70,15 +70,15 @@ class GPU_EXPORT RingBuffer {
   void DiscardBlock(void* pointer);
 
   // Gets the size of the largest free block that is available without waiting.
-  unsigned int GetLargestFreeSizeNoWaiting();
+  uint32_t GetLargestFreeSizeNoWaiting();
 
   // Gets the total size of all free blocks that are available without waiting.
-  unsigned int GetTotalFreeSizeNoWaiting();
+  uint32_t GetTotalFreeSizeNoWaiting();
 
   // Gets the size of the largest free block that can be allocated if the
   // caller can wait. Allocating a block of this size will succeed, but may
   // block.
-  unsigned int GetLargestFreeOrPendingSize() {
+  uint32_t GetLargestFreeOrPendingSize() {
     // If size_ is not a multiple of alignment_, then trying to allocate it will
     // cause us to try to allocate more than we actually can due to rounding up.
     // So, round down here.
@@ -86,9 +86,9 @@ class GPU_EXPORT RingBuffer {
   }
 
   // Total size minus usable size.
-  unsigned int GetUsedSize() { return size_ - GetLargestFreeSizeNoWaiting(); }
+  uint32_t GetUsedSize() { return size_ - GetLargestFreeSizeNoWaiting(); }
 
-  unsigned int NumUsedBlocks() const { return num_used_blocks_; }
+  uint32_t NumUsedBlocks() const { return num_used_blocks_; }
 
   // Gets a pointer to a memory block given the base memory and the offset.
   void* GetPointer(RingBuffer::Offset offset) const {
@@ -101,13 +101,13 @@ class GPU_EXPORT RingBuffer {
   }
 
   // Rounds the given size to the alignment in use.
-  unsigned int RoundToAlignment(unsigned int size) {
+  uint32_t RoundToAlignment(uint32_t size) {
     return (size + alignment_ - 1) & ~(alignment_ - 1);
   }
 
   // Shrinks the last block.  new_size must be smaller than the current size
   // and the block must still be in use in order to shrink.
-  void ShrinkLastBlock(unsigned int new_size);
+  void ShrinkLastBlock(uint32_t new_size);
 
  private:
   enum State {
@@ -117,23 +117,19 @@ class GPU_EXPORT RingBuffer {
   };
   // Book-keeping sturcture that describes a block of memory.
   struct Block {
-    Block(Offset _offset, unsigned int _size, State _state)
-        : offset(_offset),
-          size(_size),
-          token(0),
-          state(_state) {
-    }
+    Block(Offset _offset, uint32_t _size, State _state)
+        : offset(_offset), size(_size), token(0), state(_state) {}
     Offset offset;
-    unsigned int size;
-    unsigned int token;  // token to wait for.
+    uint32_t size;
+    uint32_t token;  // token to wait for.
     State state;
   };
 
   using Container = base::circular_deque<Block>;
-  using BlockIndex = unsigned int;
+  using BlockIndex = uint32_t;
 
   void FreeOldestBlock();
-  unsigned int GetLargestFreeSizeNoWaitingInternal();
+  uint32_t GetLargestFreeSizeNoWaitingInternal();
 
   CommandBufferHelper* helper_;
 
@@ -154,10 +150,10 @@ class GPU_EXPORT RingBuffer {
   Offset in_use_offset_ = 0;
 
   // Alignment for allocations.
-  unsigned int alignment_;
+  uint32_t alignment_;
 
   // Number of blocks in |blocks_| that are in the IN_USE state.
-  unsigned int num_used_blocks_ = 0;
+  uint32_t num_used_blocks_ = 0;
 
   // The physical address that corresponds to base_offset.
   void* base_;
